requirements.txt
streamlit>=1.37
pypdf>=4.3.1
python-docx>=1.1.2
openai>=1.40.0
python-dotenv>=1.0.1
docx2pdf
loguru
